public-api/requirements.txt
amqp==5.1.1
asgiref==3.5.2
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
billiard==3.6.4.0
boto3==1.23.2
botocore==1.26.2
cached-property==1.5.2
celery==5.2.7
certifi==2022.5.18
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.13
django-cors-headers==3.12.0
django-ratelimit==3.0.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
fuzzywuzzy==0.18.0
greenlet==1.1.2
haversine==2.5.1
idna==3.3
importlib-metadata==4.11.4
install==1.3.5
jmespath==1.0.0
kombu==5.2.4
mysqlclient==2.1.0
pathlib-mate==1.0.3
Pillow==9.1.1
prettytable==3.3.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pycodestyle==2.8.0
PyJWT==1.7.1
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-Levenshtein==0.12.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
SQLAlchemy==1.4.37
sqlalchemy-mate==1.4.28.3
sqlparse==0.4.2
stripe==3.3.0
toml==0.10.2
typing_extensions==4.2.0
urllib3==1.26.9
uszipcode==1.0.1
vine==5.0.0
wcwidth==0.2.5
zipp==3.8.0
web3
pillow
os